pref(group): 规范实体和模型路径
